(PC-17811) CulturalSurvey: separate utils file in single helpers
