+++ src/features/culturalSurvey/pages/CulturalSurveyQuestions.tsx
@@ -16,13 +16,13 @@ import { useCulturalSurveyQuestions } from 'features/culturalSurvey/api/useCultu
 import { CulturalSurveyCheckbox } from 'features/culturalSurvey/components/CulturalSurveyCheckbox'
 import { CulturalSurveyPageHeader } from 'features/culturalSurvey/components/layout/CulturalSurveyPageHeader'
 import { useCulturalSurveyContext } from 'features/culturalSurvey/context/CulturalSurveyContextProvider'
+import { mapQuestionIdToPageTitle } from 'features/culturalSurvey/helpers/mapQuestionIdToPageTitle'
 import {
   addSubQuestionToQuestionsToDisplay,
-  mapQuestionIdToPageTitle,
   removeSubQuestionsToDisplay,
-} from 'features/culturalSurvey/helpers/utils'
-import { useCulturalSurveyProgress } from 'features/culturalSurvey/useCulturalSurveyProgress'
-import { useGetNextQuestion } from 'features/culturalSurvey/useGetNextQuestion'
+} from 'features/culturalSurvey/helpers/questionsToDisplay'
+import { useCulturalSurveyProgress } from 'features/culturalSurvey/helpers/useCulturalSurveyProgress'
+import { useGetNextQuestion } from 'features/culturalSurvey/helpers/useGetNextQuestion'
 import { navigateToHome } from 'features/navigation/helpers'
 import {
   CulturalSurveyRootStackParamList,
